Test case for abort with error.
This test case creates an evaluation context in CPCF and an error that
should abort it, but because the PCF err-abort rule uses a
metafunction for checking that E is a non-empty *PCF* context, it
should cause a domain error.  It *should* have caused an error, but by
luck, the monitor form is considered an application in PCF.  Thus the
test passes by luck (the worst thing possible).