Label apply nodes.  Fixes 10.
Split PCF into static and dynamic variant.  Dynamic variant uses an
explicit application form `@` that is called with its source location.
This allows primitives access to source location information.  The
`err` form likewise is labelled.  This is carried through for the
other languages as well.

This eliminates the confusion of a monitor in PCF as an application
and therefore exposes #11; a test case now fails (rightly).